Added the SingleSoftware page and route. Added updateSoftware and deleteSoftware mutations. Tested and verified all Software pages and forms are working. Added context to the Hardware queries/mutations to further prevent non-logged in users from accessing any of the application data.

+++ server/schemas/resolvers.js
@@ -5,16 +5,22 @@ const {signToken} = require('../utils/auth');
 const resolvers = {
     Query: {
         //get all Hardware
-        allHardware: async (parent, args) => {
-            return Hardware.find()
+        allHardware: async (parent, args, context) => {
+            if (context.user) {
+                return Hardware.find()
                 .select('-__v')
                 .sort('type')
                 .sort('role');
+            }
+            throw new AuthenticationError('You must be logged in to view this information.');
         },
         //get single Hardware by ID
-        oneHardware: async (parent, args) => {
-            return Hardware.findOne({_id: args.hardwareId})
-                .select('-__v');
+        oneHardware: async (parent, args, context) => {
+            if (context.user) {
+                return Hardware.findOne({_id: args.hardwareId})
+                    .select('-__v');
+            }
+            throw new AuthenticationError('You must be logged in to view this information.');
         },
         //get info of logged in user
         me: async (parent, args, context) => {
@@ -23,12 +29,19 @@ const resolvers = {
                     .select('-__v');
                 return userData;
             }
-            throw new AuthenticationError('Not logged in');
+            throw new AuthenticationError('You must be logged in to view this information.');
         },
         allSoftware: async (parent, args, context) => {
             if (context.user) {
                 return Software.find()
                 .select('-__v');
+            }
+            throw new AuthenticationError('Not logged in');
+        },
+        oneSoftware: async (parent, args, context) => {
+            if (context.user) {
+                return Software.findOne({_id: args.softwareId})
+                    .select('-__v');
             }
             throw new AuthenticationError('Not logged in');
         },
@@ -63,27 +76,52 @@ const resolvers = {
             return {token, user};
         },
         //add new hardware
-        addHardware: async (parent, args) => {
-            const hardware = await Hardware.create(args);
-            return hardware;
+        addHardware: async (parent, args, context) => {
+            if (context.user) {
+                const hardware = await Hardware.create(args);
+                return hardware;
+            }
+            throw new AuthenticationError('You must be logged in to view this information.');
         },
         //update existing hardware by ID
-        updateHardware: async (parent, args) => {
-            const hardware = await Hardware.findOneAndUpdate({_id: args.hardwareId}, args, {new: true});
-            return hardware;
+        updateHardware: async (parent, args, context) => {
+            if (context.user) {
+                const hardware = await Hardware.findOneAndUpdate({_id: args.hardwareId}, args, {new: true});
+                return hardware;
+            }
+            throw new AuthenticationError('You must be logged in to view this information.');
         },
         //delete existing hardware by ID
-        deleteHardware: async (parent, args) => {
-            const hardware = await Hardware.findOneAndDelete({_id: args.hardwareId});
-            return hardware;
+        deleteHardware: async (parent, args, context) => {
+            if (context.user) {
+                const hardware = await Hardware.findOneAndDelete({_id: args.hardwareId});
+                return hardware;
+            }
+            throw new AuthenticationError('You must be logged in to view this information.');
         },
         //add a software to the database
         addSoftware: async (parent, args, context) => {
-            //if (context.user) {
+            if (context.user) {
                 const software = await Software.create(args);
                 return software
-            //}
-            //throw new AuthenticationError('Not logged in');
+            }
+            throw new AuthenticationError('You must be logged in to view this information.');
+        },
+        //update an existing software in the database
+        updateSoftware: async (parent, args, context) => {
+            if (context.user) {
+                const software = await Software.findOneAndUpdate({_id: args.softwareId}, args, {new: true});
+                return software;
+            }
+            throw new AuthenticationError('You must be logged in to view this information.');
+        },
+        //delete a software from the database
+        deleteSoftware: async (parent, args, context) => {
+            if (context.user) {
+                const software = await Software.findOneAndDelete({_id: args.softwareId});
+                return software;
+            }
+            throw new AuthenticationError('Not logged in');
         },
     }
 }
